Move emojisize and reaction enums to reside wtih message model
